PID double

+++ libraries/PID/PID.cpp
@@ -9,7 +9,7 @@ void PID::start(){
 	OldTime = micros();
 }
 
-void PID::setParameters(float _kp,float _ki, float _kd, float _pid_min, float _pid_max){
+void PID::setParameters(double _kp,double _ki, double _kd, double _pid_min, double _pid_max){
 	kp = _kp;
 	ki = _ki;
 	kd = _kd;	
@@ -17,17 +17,17 @@ void PID::setParameters(float _kp,float _ki, float _kd, float _pid_min, float _p
 	pid_max = _pid_max;
 }
 
-void PID::setParameters(float _kp, float _ki, float _kd){
+void PID::setParameters(double _kp, double _ki, double _kd){
 	kp = _kp;
 	ki = _ki;
 	kd = _kd;	
 }
 
-void PID::changeSetPoint(float _setPoint){
+void PID::changeSetPoint(double _setPoint){
 	setPoint = _setPoint;
 }
 
-void PID::limitI(float _i_min, float _i_max){
+void PID::limitI(double _i_min, double _i_max){
 	i_min = _i_min;
 	i_max = _i_max;
 }
@@ -36,11 +36,11 @@ void PID::resetI(){
 	I = 0.0;
 }
 
-float PID::Compute(float read){
+double PID::Compute(double read){
 
 	NewTime = micros();
 	dTime = ((double)(OldTime - NewTime))/1000000.0;
-	float error = setPoint - read;
+	double error = setPoint - read;
 	dError = oError-error;
 
 	P = error*kp;
