+++ libraries/PID/PID.h
@@ -5,22 +5,22 @@ class PID
 public:
 	PID();
 	void start();
-	void setParameters(float _kp,float _ki, float _kd, float _pid_min, float _pid_max);
-	void setParameters(float _kp,float _ki, float _kd);
-	float Compute(float read);
+	void setParameters(double _kp,double _ki, double _kd, double _pid_min, double _pid_max);
+	void setParameters(double _kp,double _ki, double _kd);
+	double Compute(double read);
 	void resetI();
-	float getP();
-	float getI();
-	float getD();
-	void limitI(float _i_min, float _i_max);
-	void changeSetPoint(float _setPoint);
-	float P;
-	float I;
-	float D;
-	float kp;
-	float ki;
-	float kd;
-	float _PID;
+	double getP();
+	double getI();
+	double getD();
+	void limitI(double _i_min, double _i_max);
+	void changeSetPoint(double _setPoint);
+	double P;
+	double I;
+	double D;
+	double kp;
+	double ki;
+	double kd;
+	double _PID;
 
 private:
 	
@@ -29,9 +29,9 @@ private:
 	double dTime;
 	double dError;
 	double oError;
-	float setPoint;
-	float pid_max;
-	float pid_min;
-	float i_min = 0;
-	float i_max = 0;
+	double setPoint;
+	double pid_max;
+	double pid_min;
+	double i_min = 0;
+	double i_max = 0;
 };
